Move Babysitter into a class
